Completed the sampler class using the CDF algorithm in order to sample general functions as distributions.
Checked the effectiveness of the class by integrating the functios x*exp(-x) and x*x*exp(-x) in range [0, 5] both using normal MC integration with a uniform distribution and using the sampler to
integrate the functions drawing from an exponential distribution. The test simply print the results confronting them with the real one in order to see how integrals effectivelly converges to right
results proving correctness of the method.